+++ src/ingest_validation_tools/validation_utils.py
@@ -12,7 +12,14 @@ from ingest_validation_tools.directory_validator import (
     DirectoryValidationErrors,
     validate_directory,
 )
-from ingest_validation_tools.enums import DatasetType, EntityTypes, OtherTypes, Sample
+from ingest_validation_tools.enums import (
+    OTHER_FIELDS_UNIQUE_FIELDS_MAP,
+    UNIQUE_FIELDS_MAP,
+    DatasetType,
+    EntityTypes,
+    OtherTypes,
+    Sample,
+)
 from ingest_validation_tools.schema_loader import (
     EntityTypeInfo,
     PreflightError,
@@ -20,18 +27,6 @@ from ingest_validation_tools.schema_loader import (
     get_possible_directory_schemas,
 )
 from ingest_validation_tools.table_validator import ReportType
-
-UNIQUE_FIELDS_MAP = {
-    OtherTypes.ANTIBODIES: {"antibody_rrid", "antibody_name"},
-    OtherTypes.CONTRIBUTORS: {"orcid", "orcid_id"},
-    DatasetType.DATASET: {"assay_type", "dataset_type"},
-    OtherTypes.SOURCE: {"strain_rrid"},
-    OtherTypes.ORGAN: {"organ_id"},  # Deprecated?
-    OtherTypes.SAMPLE: {"sample_id"},
-}
-OTHER_FIELDS_UNIQUE_FIELDS_MAP = {
-    k: v for k, v in UNIQUE_FIELDS_MAP.items() if not k == DatasetType.DATASET
-}
 
 
 def match_field_in_unique_fields(
@@ -86,7 +81,9 @@ def get_schema_version(
         return other_type
     message = []
     if not [field for field in UNIQUE_FIELDS_MAP[DatasetType.DATASET] if field in rows[0].keys()]:
-        message.append(f"No assay_type or dataset_type in {path}.")
+        message.append(
+            f"Required dataset field not present in {path}. One of the following is required: {', '.join(sorted(UNIQUE_FIELDS_MAP[DatasetType.DATASET]))}"
+        )
         if "channel_id" in rows[0]:
             message.append('Has "channel_id": Antibodies TSV found where metadata TSV expected.')
         elif "orcid_id" in rows[0]: